Fix several type errors across project
The return value from click's path validation is str, which needs to be cast to a path when initialising metadata.

The from_dict class method defined on DataBase had two calls using get from a dictionary which needed to default to the empty string.

The check_duplicates function was still using strings for its save_path parameter. This was updated to correctly use pathlib.Path instead.

In report_results, the files parameter was erroneously defined twice, so one was renamed.